完善接口key功能

+++ src/Framework/Leelite.Framework.Elasticsearch/Data/Store/ElasticStore.cs
@@ -1,4 +1,5 @@
 ﻿using Leelite.Core.Validation;
+using Leelite.Framework.Domain.Model;
 
 using Nest;
 
@@ -14,6 +15,11 @@ namespace Leelite.Framework.Data.Store
         /// <inheritdoc/>
         public void Add([Valid] TEntity entity)
         {
+            if (entity == null)
+            {
+                return;
+            }
+
             var response = ElasticClient.IndexDocument(entity);
             if (!response.IsValid)
             {
@@ -24,6 +30,11 @@ namespace Leelite.Framework.Data.Store
         /// <inheritdoc/>
         public async Task AddAsync([Valid] TEntity entity, CancellationToken cancellationToken = default)
         {
+            if (entity == null)
+            {
+                return;
+            }
+
             var response = await ElasticClient.IndexDocumentAsync(entity);
             if (!response.IsValid)
             {
@@ -34,6 +45,11 @@ namespace Leelite.Framework.Data.Store
         /// <inheritdoc/>
         public void AddRange([Valid] IEnumerable<TEntity> entities)
         {
+            if (entities == null || entities.Count() == 0)
+            {
+                return;
+            }
+
             var response = ElasticClient.IndexMany(entities);
             if (!response.IsValid)
             {
@@ -44,6 +60,11 @@ namespace Leelite.Framework.Data.Store
         /// <inheritdoc/>
         public async Task AddRangeAsync([Valid] IEnumerable<TEntity> entities, CancellationToken cancellationToken = default)
         {
+            if (entities == null || entities.Count() == 0)
+            {
+                return;
+            }
+
             var response = await ElasticClient.IndexManyAsync(entities);
             if (!response.IsValid)
             {
@@ -54,6 +75,11 @@ namespace Leelite.Framework.Data.Store
         /// <inheritdoc/>
         public void Remove(TEntity entity)
         {
+            if (entity == null)
+            {
+                return;
+            }
+
             var response = ElasticClient.Delete<TEntity>(entity);
             if (!response.IsValid)
             {
@@ -64,6 +90,11 @@ namespace Leelite.Framework.Data.Store
         /// <inheritdoc/>
         public async Task RemoveAsync(TEntity entity, CancellationToken cancellationToken = default)
         {
+            if (entity == null)
+            {
+                return;
+            }
+
             var response = await ElasticClient.DeleteAsync<TEntity>(entity);
             if (!response.IsValid)
             {
@@ -74,6 +105,11 @@ namespace Leelite.Framework.Data.Store
         /// <inheritdoc/>
         public void RemoveRange(IEnumerable<TEntity> entities)
         {
+            if (entities == null || entities.Count() == 0)
+            {
+                return;
+            }
+
             var response = ElasticClient.DeleteMany(entities);
             if (!response.IsValid)
             {
@@ -84,6 +120,11 @@ namespace Leelite.Framework.Data.Store
         /// <inheritdoc/>
         public async Task RemoveRangeAsync(IEnumerable<TEntity> entities, CancellationToken cancellationToken = default)
         {
+            if (entities == null || entities.Count() == 0)
+            {
+                return;
+            }
+
             var response = await ElasticClient.DeleteManyAsync(entities);
             if (!response.IsValid)
             {
@@ -94,6 +135,11 @@ namespace Leelite.Framework.Data.Store
         /// <inheritdoc/>
         public void Update([Valid] TEntity entity)
         {
+            if (entity == null)
+            {
+                return;
+            }
+
             var response = ElasticClient.Update<TEntity>(entity, p => p.Doc(entity));
             if (!response.IsValid)
             {
@@ -104,6 +150,11 @@ namespace Leelite.Framework.Data.Store
         /// <inheritdoc/>
         public async Task UpdateAsync([Valid] TEntity entity, CancellationToken cancellationToken = default)
         {
+            if (entity == null)
+            {
+                return;
+            }
+
             var response = await ElasticClient.UpdateAsync<TEntity>(entity, p => p.Doc(entity));
             if (!response.IsValid)
             {
@@ -114,6 +165,11 @@ namespace Leelite.Framework.Data.Store
         /// <inheritdoc/>
         public void UpdateRange([Valid] IEnumerable<TEntity> entities)
         {
+            if (entities == null || entities.Count() == 0)
+            {
+                return;
+            }
+
             var response = ElasticClient.IndexMany(entities);
             if (!response.IsValid)
             {
@@ -124,6 +180,11 @@ namespace Leelite.Framework.Data.Store
         /// <inheritdoc/>
         public async Task UpdateRangeAsync([Valid] IEnumerable<TEntity> entities, CancellationToken cancellationToken = default)
         {
+            if (entities == null || entities.Count() == 0)
+            {
+                return;
+            }
+
             var response = await ElasticClient.IndexManyAsync(entities);
             if (!response.IsValid)
             {
